Removed initiliazation function, initialize trough constructor
Removed the initiliaze function, and intitialize objects trough the constructor for ease.

This is because i added args to the addcomponent and still having initilliazer defeats that purpose.

Fixed all my components so that they now work with that.

Fixed small mistakes in json, and added level id parsing trough the level components